refactor: rename writeToFile -> writeToJsonFile
