+++ src/events/mineflayer/chat/promoteDemote.ts
@@ -1,7 +1,7 @@
 import Emojis from "../../../util/emojis/chatEmojis";
 import { Event } from "../../../interfaces/Event";
-import { Util } from "discord.js";
 import { HypixelRank } from "../../../interfaces/Ranks";
+import { Util } from "discord.js";
 import getRankData from "../../../util/emojis/getRankData";
 
 export default {
